+++ CHANGELOG.md
@@ -1,11 +1,18 @@
 ## Unreleased changes
 
+- The SDK requires node version 5.4 or later.
 - Support relative and absolute block height as the block identifier in block queries.
 - Add field `protocol_version` to `BlockInfo` which is the protocol version of the queried block.
 - Extend enum `PendingUpdateEffect` with variants for protocol version 6.
 - Introduce `ChainParametersV2` struct for protocol version 6.
 - Introduce generic `gas_reward` in `RewardParametersSkeleton` for supporting different versions of GASRewards.
-- The SDK requires node version 5.4 or later.
+- Add `find_account_creation` helper to find a block where the account was
+  created.
+- Deprecate `find_earliest_finalized` and replace it with
+  `find_at_lowest_height` that avoids an extra call to the node. This is making
+  use of the new API.
+- Re-export `http::Scheme` from `http` crate since it is often needed when
+  configuring endpoints.
 
 ## 2.4.0